+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Trait `StdNumOps` defines saturating math operations. All unit types, `Size`,
+  and `Point` implement this trait.
+- `Rect::saturating_extents` returns the extents of a rectangle using saturating
+  math.
+
 ## v0.4.2 (2024-10-20)
 
 ### Added